move convert functions, fix test
